backend/requirements.txt
alembic==1.13.1
blinker==1.8.2
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-SQLAlchemy==3.1.1
google-genai==1.43.0
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
numpy==1.26.4
faiss-cpu==1.7.4
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
Pillow==10.4.0
pypdf==4.3.1
pytesseract==0.3.13
pyarrow==16.1.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pytest==8.2.0
pytest-cov==5.0.0
pytest-mock==3.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.3
lightgbm==4.5.0
